requirements.txt
python-dotenv
argcomplete
requests
